- GcmIntentService für allgemeine Intents angepasst - Fehler aus GroupEndpoint behoben - ProfileGroupAdapter für Verwendung von BlobKeys angepasst

+++ SkatenightBackend/src/main/java/ws1415/SkatenightBackend/GroupEndpoint.java
@@ -849,8 +849,10 @@ public class GroupEndpoint extends SkatenightServerEndpoint {
      * können.
      */
     public void cleanPreviewPictures(){
-        for(String key : getUserGroupPreviewPictures().getBlobKeysValues()){
-            blobstoreService.delete(new BlobKey(key));
+        if(getUserGroupPreviewPictures().getBlobKeysValues() != null){
+            for(String key : getUserGroupPreviewPictures().getBlobKeysValues()){
+                blobstoreService.delete(new BlobKey(key));
+            }
         }
     }
